requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
requests
spacy
PyPDF2
python-docx
PyJWT
Werkzeug
